requirements.txt
asgiref==3.4.1
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2020.6.20
charset-normalizer==2.0.12
Django==3.2.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pip==21.2.4
pluggy==1.0.0
psycopg2==2.8.6
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.1
pytest-django==4.5.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
setuptools==58.0.4
soupsieve==2.3.2.post1
sqlparse==0.4.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
wheel==0.37.1
